requirements.txt
qrcode
Pillow
twilio
